+++ src/Editor/GUI/Widgets/GlslEditorWidgets/EWShaderComboBox/EWShaderComboBox.h
@@ -11,7 +11,7 @@ class EWShaderComboBox : public QComboBox
 {
     Q_OBJECT
 
-    QVector<unsigned long long> fullIdList;
+    QVector<uint64_t> fullIdList;
 
     ShaderComboBoxContextMenu* contextMenu;
 
@@ -21,20 +21,20 @@ public:
 
     bool eventFilter(QObject* watched, QEvent* event) override;
     void startup();
-    void selectShader(const unsigned long long key);
-    unsigned long long getSelectedId();
+    void selectShader(const uint64_t key);
+    uint64_t getSelectedId();
 
 private:
     void initialBuild(IEShaderManager* shaderManager);
     bool indexBoundCheck(const int index);
-    bool doesExist(const unsigned long long key);
+    bool doesExist(const uint64_t key);
     QString checkForDuplicateName(const QString& nameToCheck);
 
 private slots:
     void currentShaderChanged(int index);
-    void addShader(const unsigned long long key, const QString& path);
-    void removeShader(const unsigned long long key);
-    void changeShaderKey(const unsigned long long oldKey, const unsigned long long newKey);
+    void addShader(const uint64_t key, const QString& path);
+    void removeShader(const uint64_t key);
+    void changeShaderKey(const uint64_t oldKey, const uint64_t newKey);
     void showCustomContextMenu(const QPoint& pos);
 
 signals:
