+++ src/Game/Resources/FileResource/IEFileResource/IEFileResource.h
@@ -11,7 +11,7 @@ public:
     IEFileResource(const IEFileResource&) = delete;
     virtual ~IEFileResource();
 
-    unsigned long long updateId(const QString& path) override;
+    uint64_t updateId(const QString& path) override;
 
     QDataStream& serialize(QDataStream& out, const Serializable& obj) const override;
     QDataStream& deserialize(QDataStream& in, Serializable& obj) override;
