+++ src/Game/Resources/SceneResource/IESceneResource/IESceneResource.h
@@ -11,7 +11,7 @@ public:
     IESceneResource(const IESceneResource&) = delete;
     virtual ~IESceneResource();
 
-    unsigned long long updateId(const QString& path) override;
+    uint64_t updateId(const QString& path) override;
 
     QDataStream& serialize(QDataStream& out, const Serializable& obj) const override;
     QDataStream& deserialize(QDataStream& in, Serializable& obj) override;
