+++ src/Game/Scripting/API/LuaUtility/LuaUtility.cpp
@@ -20,7 +20,7 @@ void LuaUtility::addToLua(sol::table& utilityTable)
                                           "compute", &LuaUtility::computeHash);
 }
 
-unsigned long long LuaUtility::computeHash(const char* str)
+uint64_t LuaUtility::computeHash(const char* str)
 {
     // TODO this throws an exception in lua (integer value will be misrepresented in lua)
     return IEHash::Compute(str);
